feat: cleaned documents, separated into main and ants.cpp, ants.h, added timer
