+++ .idea/uneviedefourmi.iml
@@ -1,2 +1,2 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<module classpath="CIDR" type="CPP_MODULE" version="4" />
+<module classpath="CMake" type="CPP_MODULE" version="4" />
