+++ ants.cpp
@@ -1,44 +1,24 @@
-#include <iostream>
-#include <map>
-#include <vector>
-#include <queue>
-#include <string>
-#include <algorithm>
-#include <fstream>
-#include <sstream>
-
-using namespace std;
-
-class Ant {
-public:
-    string name;
-    string position;
-    bool finished;
-
-    Ant(string n) {
-        name = n;
-        position = "Sv";
-        finished = false;
-    }
-};
-
-// Structure to store ant colony information
-struct ColonyInfo {
-    int numAnts;
-    map<string, int> roomCapacity;
-    map<string, vector<string>> tunnels; // Graph of connections
-};
+#include "ants.h"
 
 // Global variables
 ColonyInfo colonyInfo;
 map<string, int> roomOccupancy;
+TimingInfo timing;
 
+// Ant class constructor
+Ant::Ant(string n) {
+    name = n;
+    position = "Sv";
+    finished = false;
+}
+
+// Add a bidirectional tunnel between two rooms
 void addTunnel(const string& a, const string& b) {
     colonyInfo.tunnels[a].push_back(b);
     colonyInfo.tunnels[b].push_back(a);
 }
 
-// Read colony file
+// Load colony from file
 bool loadColonyFromFile(const string& filename) {
     ifstream file(filename);
     if (!file.is_open()) {
@@ -47,7 +27,7 @@ bool loadColonyFromFile(const string& filename) {
     }
 
     string line;
-    
+
     // Read number of ants
     if (getline(file, line)) {
         if (line.substr(0, 2) == "f=") {
@@ -58,21 +38,21 @@ bool loadColonyFromFile(const string& filename) {
     // Read rooms and their capacities
     while (getline(file, line)) {
         if (line.empty()) continue;
-        
+
         // If line contains " - ", it's a tunnel
         if (line.find(" - ") != string::npos) {
             // Process tunnels later
             break;
         }
-        
+
         // Process rooms
         istringstream iss(line);
         string roomName;
         iss >> roomName;
-        
+
         // Default capacity = 1
         int capacity = 1;
-        
+
         // Look for capacity between { }
         size_t start = line.find("{");
         size_t end = line.find("}");
@@ -82,36 +62,36 @@ bool loadColonyFromFile(const string& filename) {
             capacityStr.erase(capacityStr.find_last_not_of(" \t") + 1);
             capacity = stoi(capacityStr);
         }
-        
+
         colonyInfo.roomCapacity[roomName] = capacity;
     }
 
     // Process current line (first tunnel) and following ones
     do {
         if (line.empty()) continue;
-        
+
         size_t pos = line.find(" - ");
         if (pos != string::npos) {
             string room1 = line.substr(0, pos);
             string room2 = line.substr(pos + 3);
-            
+
             // Clean spaces
             room1.erase(0, room1.find_first_not_of(" \t"));
             room1.erase(room1.find_last_not_of(" \t") + 1);
             room2.erase(0, room2.find_first_not_of(" \t"));
             room2.erase(room2.find_last_not_of(" \t") + 1);
-            
+
             addTunnel(room1, room2);
         }
     } while (getline(file, line));
 
     file.close();
-    
+
     // Initialize room occupancy
     for (const auto& room : colonyInfo.roomCapacity) {
         roomOccupancy[room.first] = 0;
     }
-    
+
     return true;
 }
 
@@ -200,9 +180,10 @@ bool compareRooms(const pair<string, int>& a, const pair<string, int>& b) {
         int numB = stoi(roomB.substr(1));
         return numA < numB;
     }
+    return roomA < roomB;
 }
 
-// Presentation for results
+// Presentation of results
 void printColonyInfo() {
     cout << "+++ Ant Colony Information +++" << endl;
 
@@ -232,107 +213,33 @@ void printColonyInfo() {
     cout << endl;
 }
 
-// Main program loop
-int main(int argc, char* argv[]) {
-    string filename;
-    
-    if (argc > 1) {
-        filename = argv[1];
-    } else {
-        cout << "Enter the ant colony filename: ";
-        cin >> filename;
-    }
+// Timing functions
+void startLoadTimer() {
+    timing.start_load = clock();
+}
 
-    if (!loadColonyFromFile(filename)) {
-        return 1;
-    }
+void endLoadTimer() {
+    timing.end_load = clock();
+}
 
-    printColonyInfo();
+void startSimulationTimer() {
+    timing.start_simulation = clock();
+}
 
-    // Create ants
-    vector<Ant> ants;
-    for (int i = 1; i <= colonyInfo.numAnts; ++i) {
-        ants.push_back(Ant("f" + to_string(i)));
-    }
+void endSimulationTimer() {
+    timing.end_simulation = clock();
+    timing.end_total = clock();
+}
 
-    cout << "Starting simulation with " << ants.size() << " ants" << endl;
-    cout << endl;
+void printTimingStatistics(int totalSteps) {
+    double load_time = double(timing.end_load - timing.start_load) / CLOCKS_PER_SEC;
+    double simulation_time = double(timing.end_simulation - timing.start_simulation) / CLOCKS_PER_SEC;
+    double total_time = double(timing.end_total - timing.start_total) / CLOCKS_PER_SEC;
 
-    int step = 1;
-    bool allFinished = false;
-
-    while (!allFinished && step <= 50) { // Safety limit
-        allFinished = true;
-
-        vector<pair<int, string>> plannedMoves;
-        map<string, int> tempOccupancy = roomOccupancy;
-
-        // Sort ants by priority
-        vector<int> antOrder;
-        for (int i = 0; i < ants.size(); i++) {
-            if (!ants[i].finished) {
-                antOrder.push_back(i);
-            }
-        }
-
-        // Prioritize ants closest to Sd
-        sort(antOrder.begin(), antOrder.end(), [&](int a, int b) {
-            vector<string> pathA = findShortestPath(ants[a].position, "Sd");
-            vector<string> pathB = findShortestPath(ants[b].position, "Sd");
-
-            if (pathA.size() == pathB.size()) {
-                return a < b; // f1 before f2, etc.
-            }
-            return pathA.size() < pathB.size(); // Closest first
-        });
-
-        for (int antIndex : antOrder) {
-            Ant& ant = ants[antIndex];
-            string current = ant.position;
-
-            string nextRoom = chooseBestNextRoom(current, tempOccupancy);
-
-            if (!nextRoom.empty()) {
-                plannedMoves.push_back({antIndex, nextRoom});
-
-                // Update temporary occupancy
-                if (current != "Sv" && current != "Sd") {
-                    tempOccupancy[current]--;
-                }
-                if (nextRoom != "Sv" && nextRoom != "Sd") {
-                    tempOccupancy[nextRoom]++;
-                }
-            }
-
-            if (ant.position != "Sd") allFinished = false;
-        }
-
-        // Apply planned moves
-        if (!plannedMoves.empty()) {
-            cout << "+++ Step " << step << " +++" << endl;
-
-            for (auto& move : plannedMoves) {
-                int idx = move.first;
-                string from = ants[idx].position;
-                string to = move.second;
-
-                // Update real occupancy
-                if (from != "Sv" && from != "Sd") roomOccupancy[from]--;
-                if (to != "Sv" && to != "Sd") roomOccupancy[to]++;
-
-                ants[idx].position = to;
-                if (to == "Sd") ants[idx].finished = true;
-
-                cout << ants[idx].name << " - " << from << " - " << to << endl;
-            }
-
-            cout << endl;
-            step++;
-        } else {
-            break;
-        }
-    }
-
-    cout << "All ants have reached Sd in " << step - 1 << " steps!" << endl;
-    return 0;
+    cout << "+++ Performance Statistics +++" << endl;
+    cout << fixed << setprecision(6);
+    cout << "File loading time: " << load_time << " seconds" << endl;
+    cout << "Simulation time: " << simulation_time << " seconds" << endl;
+    cout << "Total execution time: " << total_time << " seconds" << endl;
+    cout << "Average time per step: " << (totalSteps > 0 ? simulation_time / totalSteps : 0) << " seconds" << endl;
 }
